requirements.txt
Django==2.0.4
Pillow==5.1.0

social-auth-backend-epita==0.0.4
social-auth-app-django==2.1.0
python-social-auth[django]

reportlab==3.4.0

paypal==1.2.5
django-paypal==0.5.0

django-tastypie
